backend/pyproject.toml
[project]
name = "socratictutor"
version = "1.0.0"
description = "Socratic AI tutor for technical courses"
requires-python = ">=3.11"
license = "Apache-2.0"
authors = [
    { name = "Pratik Mehta" },
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.6",
    "sqlmodel>=0.0.21",
    "openai>=1.30.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt<4.2.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.2",
    "cuid2>=2.0.1",
    "pydantic-settings>=2.5.2",
    "email-validator>=2.0.0",
    "pyyaml>=6.0",
    "packaging>=24.0",
    "html2text>=2024.2.26",
    "python-multipart>=0.0.22",
    "trafilatura>=2.0.0",
    "beautifulsoup4>=4.12",
    "youtube-transcript-api>=1.2.4",
    "playwright>=1.58.0",
    "alembic>=1.15.0",
]

[project.scripts]
seed = "seed:seed"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "llm: requires live LLM API keys (deselect with '-m not llm')",
    "wiki: requires local pedagogy-wiki content on disk",
]
addopts = "-m 'not llm and not wiki'"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
